[FL-30032] prohibit generate-method completionItems
On the backend it's implemented through CompletionPreselectSkipper, not to pollute the public API and not to use any shady decoupled logic.
Even though initially the Skipper mechanism was introduced into completion to avoid placing the undesired element on top of the lookup; Fleet uses the same extensions
to completely eliminate elements from the lookup.

GitOrigin-RevId: cfe54281d432fee1ca28b1a56edd9cf90f510ce3