KTIJ-32220 [kotlin] Make `PublicApiImplicitTypeInspection` in K2 Mode report private declarations
This brings the behavior of this inspection to
the intended one, which was present in K1 Mode before
an accidental bug was introduced to the
`ExplicitApiDeclarationChecker.returnTypeRequired`
function from the K1 Frontend.

This change intentionally only compares
visibilities strictly by `==`, and not
by "less/more visible" relations.
This would probably be changed as a part of KTIJ-32219,
if we are going to introduce a singe selector for
the `PublicApiImplicitTypeInspection` inspection's lowest
visibility settings.

^KTIJ-32220 Fixed

GitOrigin-RevId: 1fdea25144e19734d6c9931c6bdc282752f83837