PY-77463: NPW: Disable "OK" until SDK is set.
Assume we have the following sequence of events:
1. SDK list is empty
2. `setItems` gets called
3. SDK list set its status to "busy"
4. It gets filled with interpreters and first one is chosen
5. "Busy" flag is removed

For validation, we used to check "Busy" flag only: when it was "busy" we disabled "OK" button.

But before step 3 it isn't busy nor it has interpreter set. Clicking "OK" at this moment might lead to NPE (see the issue).

We now check if value is set fore validation.

GitOrigin-RevId: 99554914cb7dc67158828559ba2e6da81fdfe350