Ensure that `TaskInfoEntity` is removed if an exception happens during creation
The task might get leaked (never removed from DB) if the coroutine is canceled during task creation.
In this case, `CancellationException` will be thrown on `withContext` exit inside `withKernel` and code that suppose to remove the task (`PlatformTaskSupport.withBackgroundProgressInternal`) won't be executed.

To ensure that the task is removed, I add try/catch to the task creation code and remove a task from there if it was created

GitOrigin-RevId: 1b67b4c80b56a1dc3e57bcec3efd496b23c7822e