IJPL-171111 Fix EntityDoesNotExistException
The problem was that the entity is created and accessed in different coroutine scopes.
Even though we guarantee the order of events and even have the created entity on hand, the second scope might not have received the updated DB yet.

To ensure that entity exists on subscription, we should not use `cs.launch`.
We should subscribe in the same scope under simple `launch`

GitOrigin-RevId: 1495d22948979a62cc9c275955d8e8fcbb5d2e5b