[platform API] remove @ApiStatus.Internal from OptionalReformatActions (IJSDK-2292)
It was marked as internal in 9f89109681d1c68, but it appears that some plugins use ReformatCodeAction.reformatDirectory, which accepts an interface extending OptionalReformatActions. A proper API should be written instead, which takes a builder for formatting options instead of requiring to implement these interfaces, but until it's done, it's ok to use OptionalReformatActions directly.

GitOrigin-RevId: da63da8e791643d96f0f12f20e164fe32542de0c