IJPL-172285 Focus new build view if the toolwindow is focused
When the builds view removes a view, the focused component disappears.
When a new view is added, it doesn't become focused automatically.
This leads to a weird state when the tool window is "sort of, but not technically"
focused. In the Dock Unpinned mode, it means that when it loses the focus,
it isn't hidden because technically it didn't have focus.

Fix by tracking the focused state of the entire view. If the focused component
is in the view, mark the view as focused. Then, when a new view is added
and made visible, make it request focus if the window is focused.
This way we ensure that a new view is always focused if the parent view
was focused before the view was removed. If the entire view is unfocused
(say, focus went somewhere else entirely), then we won't do anything
and therefore won't steal focus.

GitOrigin-RevId: fd9768ea6e1d8ba1538276b17a591e6895fd4f18