IJPL-161009 Don't auto-hide unpinned tool window on action from it
Hiding unfocused unpinned tool windows on any action seems like
a good idea. Similarly we hide tooltips, for example.
But when the action in question was invoked from the tool window
itself, it doesn't look good. Clicking a button inside a tool window
should not hide it, whether it's focused or not.

Fix by adding a predicate to the hiding function and checking
whether the tool window ID matches the tool window ID
of the component that triggered the action.

GitOrigin-RevId: 26fed84839051a25d9ad314638bf24b444f0909b