IJPL-166201 Fix tool window header side component height
The tool window header is a bit unusual: instead of using a regular
border, it paints the border as needed on top of itself.
The top/bottom borders appear and disappear depending
on a lot of heuristics, so it's not easy to account for them
when laying out the header components. Since side components
aren't used that often at all (it's an internal thing and used
so far only for the Services tool window), the easiest fix
is to simply add 1px gaps above and below the side component.

GitOrigin-RevId: 3fc3008aaba488a17b2607b2281bd479f9fc8f96