PY-77583 Properly detect an interpreter version of a uv-based venv
PythonSdkFlavor#getName is expected to return the prefix of Sdk#getVersionString result,
which is normally an output of `/path/to/python -V` or a similar command
(see the implementation of PythonSdkFlavor.getLanguageLevelFromVersionString).
Since uv primarily manages CPython-based venvs, its getName() for them should
be "Python", not "Uv". PythonSdkFlavors for Pipenv, Poetry, and venv achieve that
by extending CPythonSdkFlavor.

Because an incorrect prefix was returned, we defaulted to LanguageLevel.getLatest()
that was then set for project files by PythonLanguageLevelPusher despite the right
version string recorded in SDK data.

However, one can also install pypy interpreters with uv. We should think of a way to
properly recognize their versions as well.

GitOrigin-RevId: c7331653c34fc5814d2de08318cd44462a96eace