IJPL-171509 Introduce PopupShowOptions
To add some fine-tuning to how and where a popup is shown
without changing its API every time, let's add an options
interface.

This interface has a bit unorthodox implementation of
the Builder pattern. It's a builder itself, but then it's
used to build the actual options. This can be a bit confusing,
as the options don't implement the options interface,
but it's OK as we don't expose these details to the clients.
This is done so that various show() overloads can actually
change the options to their own liking.
All these show() overloads now converge at showImpl(),
which then builds the final options and uses them to
do everything.

Existing show() overrides were changed to override
showImpl() instead.

GitOrigin-RevId: 56e0c2430cf82dc4db24adabbdb16fc6911424d0