[RDCT] IJPL-166008: Support remote configurable groups
The core issue with Database settings was that "Database" is not a configurable, but a configurable group registered as `ConfigurableGroupEP`. Those were not supported in Remote Development, so all the real database-related configurables where transferred from the backend, but since there were no suitable parent they were missing from the tree in the settings dialog.
This commit adds such groups as stub configurables, allowing them to be a suitable parent for other remote configurables.

GitOrigin-RevId: 273e3e032bc2f6f8b84394595ee9f0a4a8c3764d