[collab] Don't update text if it's the same text (IJPL-172459)
#IJPL-172459 Fixed

This solves an issue where the caret will always jump to the end of the text.

In this particular case, there's a loop between JBTextArea reporting its value and a VM updating the JBTextArea text.
By checking the text for equality, we detect and close the loop.

I believe this should not affect other uses of bindTextIn,
since it shouldn't be the case that we rely on bindTextIn to perform this caret-reset.


(cherry picked from commit 420ed54de52fdc8937abdfb4f566dd0b4085f101)

IJ-CR-150511

GitOrigin-RevId: 67fdd6b0b354dfc4a3f387355daad45e5e6e66b6