[remote dev] restart the frontend process only calling 'restart' (IJPL-165387)
There is no need to ask about stopping the backend process when the frontend is restarting. So here a new method ApplicationListener.canRestartApplication is introduced, and it's overriden in the frontend process to suppress the confirmation before restarting.


(cherry picked from commit d68298df0513978a046829ec87fa1134ceefa1ca)

IJ-CR-150411

GitOrigin-RevId: eab235b69ca2252fe0170870a8f856f2a259407b