IJPL-171509 Implement JBPopup.showAboveOf
It turned to be more complicated than it would seem.

First, the logic of adjusting the popup position
depends heavily on its relative position to the owner
and various anchors.

Second, if its size is reduced due to minimumHeight,
then we can only apply it after creation, because
it seems to ignore the size otherwise and stick
to the preferred size anyway.
This is sort of in line with how it was implemented in VcsUIUtil,
as it "intercepts" the size before the popup is shown,
but after it's created.

GitOrigin-RevId: 7184254f27d45d1641e9299a0a26e91aff5f5db1