[github] Fix URL encoding of query params (IJPL-72781)
#IJPL-72781 Fixed

This commit removes old methods of building the query string and
always uses the DSL-way of creating the query string.
This is basically only to guarantee parameters are URL-encoded...

Removes the reliance on `toString` implementations.
`toString` really shouldn't be used in any other way than for a display string meant for developer's eyes only IMO.
Didn't remove `GithubRequestPagination.toString` because of external usages, but would like to nullify this use as well.

Also URL-encodes query parameter names, but maybe this is a bit overkill.
I have no proven case for this...

GitOrigin-RevId: 72dbbd52add0590e57d727283214b9c9f8be530e