IJPL-43545 Add IDE main frame small size detection
Since we haven't been able to figure out the root cause
of the issue when the IDE frame suddenly becomes
way too small, let's at least add some warnings here
and there if the size is unreasonably small.
Maybe we can at least pinpoint the place in the code
where it happens.

GitOrigin-RevId: 153874269a0819bf14e3b411b30aa9b5973ab3d6