IJPL-172583 Rename redispatchIfTrackNotClickable
Even if the track is not clickable, we sometimes need
to process events. For example, clicking on the thumb
should always work, as the thumb is always visible.

This is just a refactoring, the first step:
we rename this function to better reflect what it
should do: pass events only through invisible parts.

GitOrigin-RevId: d6990f254c2054876aeb650216821c681272fc62