IJPL-172583 Always process scroll bar thumb mouse pressed events
Regardless of whether the scroll bar thumb is visible or not,
mouse pressed events on it should always be processed.
First of all, in most themes / color schemes the thumb is always visible,
even if smaller or partially transparent. Except for special cases
like the editor tabs scroll bar.
Besides that, when the mouse is over the thumb, it SHOULD
be visible regardless of the theme / color scheme.
If it's not visible or partially visible, that's a bug, and such
bugs shouldn't affect the user's ability to actually use the thumb.

GitOrigin-RevId: 618f4545a78cf8bcc835fa2e5ac06520206622b2