IJPL-172583 Fix scroll bar animation
When a scroll bar is requested to appear or disappear
there are three cases:

1. It's in the exact opposite state: in this case the animation
is scheduled with a delay. This is done in case the mouse
moves past very quickly, so then the scroll bar won't flicker.

2. It's in an intermediate state. This means some animation
was started, but now we change our mind and need to
switch the direction ASAP. In this case the request
is processed without any delays.

3. The scroll bar is already in the desired state.
In this case we shouldn't really do anything,
so it was coded to do exactly this.
But it was a bug because there might be a pending
request waiting to start. We need to cancel it
to ensure that the scroll bar stays in the desired state.

To fix, make the animation requests nullable
with null corresponding to "do nothing".
Emitting a null makes collectLatest cancel
the previous request. It's guaranteed to cancel
before it actually starts executing because
both cancellation and request execute on the EDT.

GitOrigin-RevId: 8b23fce4d01877e09905946c1b954e609472bcd9