[java-dfa] IDEA-364131 Inline overridable trivial accessors if the exact qualifier type is known
Getter inlining functionality moved to MethodCallInstruction

Also: get rid of MethodCallInstruction.myPrecalculatedReturnValue. In any case, we do a lot of ad-hoc computations inside MethodCallInstruction. So we can move two cases when myPrecalculatedReturnValue was used (getter processing and empty collection processing) into MethodCallInstruction as well. This unifies code a little, and provides more possibilities, as we know current abstract interpretation state and can use it (in particular, stability of qualifier)

Also: a new kind of MutationSignature 'transparent', which doesn't flush even private fields. It's not exposed to public annotations yet (appears to the user in the same way as 'pure')

GitOrigin-RevId: 548ab12afc0d0314829f47e0ef51caec59985698