IDEA-351682 Fix immediate closing of the debugger if run configuration has a specific name
This listener was added to support removing of the top level tool window tab if the corresponding inner debugger tab was hidden in the settings.
It is specific to the case when inner debugger tabs ('Console', 'Threads' and so on) are shown as a top level tool window tabs (`debugger.new.tool.window.layout.single.content` registry option).
It is implemented in a very inaccurate way and may cause such problems as IDEA-351682.
I decided to remove it since single content layout is not used now, and anyway, the fix for the original problem should be implemented differently.

GitOrigin-RevId: aaa140c22a7c07ea25df67b1df1ffa79dd2fd9fd