[platform API] cleanup: move mac.screenmenu package to impl module
These classes are used from the implementation only, so there is no reason to have them in the API module. Also, this ensures that the reference to ActionsBundle is resolved (corresponding ActionsBundle.properties is located in intellij.platform.resources.en module which isn't included in the dependencies of intellij.platform.ide module).

GitOrigin-RevId: dad7e2b2e2b631212b3a69f85089785006a64075