IJPL-49322 Speed search customization in AbstractPopup
Make setSpeedSearchAlwaysShown public
and add a method to set the placeholder text.

This is needed for the new file creation popup in the Project View,
to avoid subclassing the popup like it's done for the Branches popup.

The assertion for setSpeedSearchAlwaysShown is relaxed
to allow setting it in the INIT state or in the NEW state.
It's not necessary to call init() before setting a simple
variable, the most important part is to set it before
it's used. The previous version of the assertion broke
some RemDev tests, as they use popups in the NEW state.

GitOrigin-RevId: 6f3caf768b178896feb9547a65b6186067b26457