IJPL-172936 Only apply the calculated popup size if it was changed
There's a regression after a recent change: if the popup has a minimum
height set, its calculated size is applied to it, ignoring whatever
size the popup sets for itself. Ideally, these sizes should be the same
if there are no adjustments applied, but unfortunately, it's not always
the case, not clear why. Likely it's related to some insets or some
adjustments that the popup may make itself.

Trying to reason about this mess is risky and fragile. Even if we find
the cause now and fix it, something else can go wrong. Ideally,
we shouldn't tell the popup its size unless it's really necessary.
So the fix is to only do it if we actually applied size adjustments,
not when the popup simply happens to have a minimum height.

Introduce a private record for these adjustments. We could do
with a single Boolean value now, of course, but having a record
makes the code more readable and allows us to extend it in the future.

GitOrigin-RevId: a837d25caecaaff805c731cc9fc71ab9da07105e