IJPL-60835 Only allow to customize left/right/center of the main toolbar
Toolbar customization shows a tree of actions for its root,
which contains the left, center and right groups. However,
actual customization is only possible for the contents
of these groups. Moving, deleting and adding groups at the top
level isn't supported.

Fix by extending CustomizableActionsPanel with
the minSelectionPathLength() method. There's already
a similar one in its private actions, and they use it
to determine which actions are possible.
For the root it's possible to add actions,
below that level all other actions (edit/delete/move)
are also possible.

By overriding this value we specify for the main
toolbar that only 2nd level groups should be treated
as customization roots, thus effectively disabling
the Add action for the top level and various
edit actions for the 2nd level.

GitOrigin-RevId: ce72e18ebdae18ee1ecd90aef579638ac0864505