IJPL-171509 Convert JBPopup.showAboveOp to show(options)
Because everything we need is now represented in the options,
adding a separate showAboveOf method doesn't sound like
a good idea anymore.

Let's try to introduce a single entry point show(PopupShowOptions).

In BackendBeAbstractPopup, use the passed options
to calculate the anchors, if they're provided.

Instead of having one general factory function for the options,
introduce popupAboveComponent with clear intent
and a mandatory argument. We can later do the same
for other use cases, providing a nice set of default
options right out of the box, with the ability to modify
anything that the caller might want to modify.

GitOrigin-RevId: 9ddf263f85dc1cef0b4c80d06ff7864be0064ea0