IJPL-173127 If a popup doesn't fit above, try to show it below
The new functionality in AbstractPopup that allows
showing the popup above the owner doesn't work well
if there's no space even for the reduced height.
This often happens when there's plenty of space below,
so it's best to show it below if it doesn't fit.

Fix by reverting the move above the owner
if the component doesn't fit there.
Then the code that tries to fit it below will kick in.

GitOrigin-RevId: 193f75ec85910dce4816228c1145d090d01ea4a1