IJPL-49322 Disable the New Element action if there's no selection in the Project View
Now that we intend to have an action button for this action,
it'll be very confusing when the user starts the IDE,
clicks this button without any selection and get an almost empty
popup offering to create a scratch file.

The no-selection state is very temporary in its nature anyway,
as once something is selected, the selection normally doesn't go away.
So the action will remain disabled until the user selects something,
and then will remain to be enabled.

GitOrigin-RevId: 4cfdfbfed25c6e8a40366cebdb15f28e974b63b7