IJPL-49322 Implement the customized empty text for New Element
Implementing the custom empty text is easy, as we already have
the necessary mechanism.

Actually invoking the action from it is harder, though,
as the mechanism doesn't support IJ actions, only Swing actions.
There has to be a better way of doing this, but for now
this ad-hoc approach of creating stuff manually
and invoking the action by ID seems to work OK.

GitOrigin-RevId: dd1503f4b3b5c478fe67b18d5282933bb69defc3