[build scripts] refactoring: allow using ProductInfoData for deserialization of product-info.json from previous versions
Properties which weren't required in previous versions are marked as nullable with default versions provided in the primary constructor to allow using it for deserialization of product-info.json files from previous versions (see IJPL-356). Separate factory functions are introduced with parameters for these properties having not-null types to ensure that the build scripts always generate values for them.

'arch' is removed from the list of required properties in product-info.schema.json because it wasn't present in old versions.

GitOrigin-RevId: be222c9efc1450f8524610bee456dcab34cf69a1