KTIJ-32349 [kotlin] Use original file `analyze` in `OptimizedImportsBuilder` to analyze original references
This is done to avoid
"The element cannot be analyzed in the context of the current session"
error (see KT-66783).

It happens in J2K because the file that comes into the import optimizer
is actually dangling.
And currently there is no way to analyze two dangling files
simultaneously (see KT-73836).

We used to do that to compare the resolve with old and new imports.

Now we use the original (potentially dangling file) to do the analysis
of the original reference, save the `SymbolInfo`s (which do not contain
`KaSymbol`s now), and compare them without relying on `KaSession` at all.

Add new test to the J2K suite, which triggered the exception before.

^KTIJ-32349 Fixed

GitOrigin-RevId: 665ba226638211612dd0c5ea1e48aa9dedc30bb4