IJPL-188144 Introduce inspection for checked exceptions in Kotlin
This inspection can work only in the IntelliJ monorepository. It's because the author has doubts that the overall idea of having that inspection is good.

The inspection was introduced for Eel Api as a replacement for the Either monad. It's convenient enough to work with `Result` in Rust, but we didn't manage to make a convenient analog for Kotlin.

Differences from Java's checked exceptions:
* This inspection produces warnings and doesn't break compilation.
* By design, it's fine to suppress the exception.
* The inspection should not try to prove in every case that some exception is not caught. It's supposed only to remind API users about exceptions.

GitOrigin-RevId: 957d2ff982c7ec5f52da72d7027306615827a1c7