PY-73050 Remove hardcoded return types for open() functions
Now that we have proper support for Literal types and overloads,
we can infer the right return type automatically.

It breaks custom type inference for the "open" function in Python 2
because of their outdated stubs not using Literal type for the "mode"
argument. The stubs for Python 2 are no longer included in Typeshed,
so we keep an old snapshot of them. However, breaking it seems
acceptable nowadays, and if the need ever arises, we should better
customize our copy of .pyi stubs for Python 2.

GitOrigin-RevId: cd18252976959e81e2412f0433edc66e8e9ccfc5