IJPL-158493 Fix presentation update bugs in the tree structure adapter
Add an explicit node descriptor update there. This means
that in some cases the descriptor will be updated twice.
Unfortunately, due to the asynchronous nature of the new
model there's no simple solution for this, like in the old
implementation when we could simply build a set of freshly
created nodes and not update them because they're updated
on creation. Here, for now, we'll have to live with double
updates.

Port the color computation mess from NodeRenderer as well.
This fixes VCS colors, for instance.

GitOrigin-RevId: 9287e3c40ab879bcf9ac9e9f12131c0348216c1a