IJPL-173577 Replace ActionGroup.addAll with add in tool windows
The addAll method is deprecated as it uses getChildren(null).
We don't really need to expand the group here, as the menu will do it
anyway. And if the group isn't a popup group (which it doesn't seem to be),
then it'll be expanded on the same level, with the same effect
as with adding children manually.

GitOrigin-RevId: 9bfc38651772d678daa652573dec1a891c14c1e5