Revert "IJPL-60835 Only allow to customize left/right/center of the main toolbar"
This reverts commit ce72e18ebdae18ee1ecd90aef579638ac0864505.

It's reverted because this approach only worked for specific dialogs,
e.g., the main toolbar customization dialog. But for the generic customization
configurable in the Settings it doesn't work at all. It contains a lot of stuff,
and customization must be only restricted for some subtrees, which can't
be expressed via the "minimum path length" concept. We need a better
solution.

GitOrigin-RevId: 79301a9ea99f14bb161e59d5c1e501ed3eb59238