IJPL-60835 Refactor CustomizableActionsPanel actions for BGT update
We'll need some additional logic to disable actions for certain
action groups, like the main toolbar left/center/right groups.
This logic will likely involve some extensions to avoid hardcoding
action groups here. Extension loading is best done on a BGT,
so let's refactor these actions first to use the data context
instead of accessing the UI directly from update().

In order for it to work, we need a custom UI provider panel here,
and we also need to set it to be the target component of both
toolbars. That the top panel was the target component of
one of the toolbars was almost surely a bug.

GitOrigin-RevId: 7c3cb7965e7472eda164b7535726543691773817