IJPL-60835 Introduce ActionGroupCustomizationService
To be able to specify which action groups can be customized,
we drop the idea of the minimum selection length and introduce
the concept of a customization root.

By default, the top-level action groups are considered
customization roots. They have the path length of 2
(including the invisible root), and we can add elements to them,
but we can't rearrange or edit them. For all elements below
these, all actions are allowed. This corresponds to
the previously existing selection length checks.

Then we redefine a customization root as some action group
that is under a read-only parent. By default, this read-only
parent is the invisible root. But we also introduce a service
to be able to specify additional read-only action group IDs.
Then we can extend this service so that specific action
groups may be added anywhere.

GitOrigin-RevId: 54ff4fdb5ed4276d0ce1cfa8869dce4ec0b7909c