[Gradle KMP] Changed the way of collecting source sets for jar artifacts.
Previously, we collected compilations by iterating via the target's list of tasks on which an archive task was dependent.
With kotlin 2.1.20, we added an archive task to the configuration.
So with an enabled isolated projects feature we don't need to do the old logic, but we still need to support the old logic for the older kotlin versions.

^KTIJ-32183 Fixed

GitOrigin-RevId: 81a283b31e9c8c9bde3fef4ef895f234452a6719