[platform] Optimize searching for the central directory in JBZipFile
In the default implementaiton of JBZipFile, the central directory is read byte-by-byte. This is not convenient for remote setup, where each read of a byte is a separate RPC. Moreover, if the archive is malformed, then the implementation would issue an RPC for each byte of the archive, which could be quite large. That leads to unacceptable slowdowns, which can be prevented by preventive caching of the data that we are going to read.

GitOrigin-RevId: 658070e34a238fdf86d9d183f011c36ce48be915