IJPL-61637 Fix ToolbarComboButton height in the main toolbar
The height is determined by the toolbar layout, which enforces
all components to be at least as high as
experimentalToolbarMinimumButtonSize, but it was decreased
to be less than the toolbar height because for action buttons
it doesn't include their insets. So for example when the toolbar
is 40 px high, action buttons are 30x30 and the height is set to 30,
not to 40.

For ToolbarComboButton it means that the hover area has the right
height, which is 30, but the size of the button itself is smaller than
the header. Therefore, one can't click such a component in a maximized
window when the mouse is at the top. This is, of course, bad UX
in systems such as Windows, and it's inconsistent with other components.

The fix is similar to what we do with action buttons:
we specify the height to the toolbar combo button,
as a supplier to account for theme changes without having
to react in updateUI. To mirror action button behavior,
this size doesn't include the insets.

The insets themselves are calculated from
mainToolbarButtonInsets, but without the left/right ones,
as we're only interested in the height here.

Now, for example, if the toolbar is 40 px high,
we'll have the hover area of 30 px plus 10 px worth of insets.
This both makes the button clickable through the entire
height of the toolbar and preserves the height of the hover area.

GitOrigin-RevId: d8a5337937dd9dd46ca86f48d80b06d0898d29ff