Eel: add `stderr->stdout` redirection part 1: Add Eel API.
These patches enable `stderr->stdout` redirection much like `redirectErrorStream` in JVM.
For the local implementation we will use JVM, for ijent-based `dup2(1,2)` for **nix and `hStdErr = hStdOut` for Windows will be used (Win part will be implemented some day).

This change introduces eel exec API part for it.

There are three possible states:
* No pty nor redirection.
* StdErr redirected to StdOut, but still no pty.
* Pty (which implicitly assumes stderr is redirected to stdout. Both streams are redirected to the tty actually).

This change reflects it in the API.

GitOrigin-RevId: c07f2e9bbaecc8e667659ee843f480092689711a