Eel and Ijent TTY: Set `TERM` if PTY for local processes, add test and doc.
When running under the tty, a process should always have a `TERM` variable. See `com.intellij.platform.eel.EelExecApi.ExecuteProcessOptions.Builder.pty` doc.

If there is no one in `env`, we set it to `xterm`.

This contract was implemented by ijent.

We document it, cover with tests and implement for local eel

GitOrigin-RevId: 07e3999c1cd561d02d86ebf53fab9d70ffca358b