Python tests: Add pythons to the list of allowed roots.
In tests, we check that no code access files outside so-called roots. Roots are usually project roots, but we might need to access files in python installation dirs (to access `black.exe`, `poetry.exe` and so on)

We already fixed it for JUnit5 tests. We now fix it for the legacy `PyEnvTestCase`

See `VfsRootAccess`.

GitOrigin-RevId: 494e60d3d07b9d14b4001b7f37f4f5188d63c994