[platform api] remove the 'internal' mark from 'productInfo' API
This change adds serializable classes describing schema of the product-info.json file to the API dump. This is needed to make changes in the format of these files visible, it's important, because we have quite a few tools which rely on them (e.g., Toolbox, IntelliJ Platform Gradle Plugin, Gateway).

GitOrigin-RevId: 931bd21b44f0dd8f746f2ec64f626c87ef1bc0d6