CPP-42585 more precise check for editor window shown in a DockWindow
DockManager.getContainerFor() returns DockManager.dockContainer for editors
opened in the main window. This made us think that the singleton editor opened
in a Split is shown in a DockWindow. As a result the same singleton file
with disassembly was opened in another Split in the main window.


(cherry picked from commit 8584fd131615209cf26ff70315d959828a6a46d2)

IJ-CR-151729

GitOrigin-RevId: 201e364498109091c37d35cb43c59014eb90f58f