IJPL-171238 Get rid of `CoroutineScope` in `TaskSuspender`
The scope now will be passed in `PlatformTaskSupport` when the task is started under progress.
This allows the suspender to be created from blocking context, and there is no need to control the scope from the outside - it's going to be controlled by `PlatformTaskSupport`.

However, this means that the suspender won't be fully workable until the task is started (it won't receive updates from `CoroutineSuspender` until then)

GitOrigin-RevId: 0d4898a716974a69801394cd3a937e88176f4a7e