[remote dev] rename JARs and directories in CWM plugin for consistency (IJPL-170362)
Names of JARs should be consistent with names of modules packed in them to avoid confusion. 'client' directory is renamed to 'frontend-split' to be consistent with the actual naming schema. Consistent names are also needed to avoid changes in the layout when the JAR names for modules used by the split frontend process will be chosen automatically (IJPL-170362).

GitOrigin-RevId: 4e76878d04d6335be5436f812084c4ff73f38540