IJPL-160390 Add `Source` to `TaskStatus` events to distinguish events which should be processed by `PlatformTaskSupport`
Changes in `TaskSuspender` generate changes in `TaskStatus`, so in `PlatformTaskSupport.subscribeToTaskStatus` each change is going to be processed twice.

For example, if `TaskSuspender` becomes paused, we call `TaskManager.pauseTask` which changes the status of the task in DB.
After that we receive task status change and call `taskSuspender.pause`.

It won't cause any problems if events happen with a delay (Paused, delay, Resumed).
But if they happen instantly we might get unexpected state:
1. `TaskSuspender.pause` -> suspender paused
2. task status is set to `Paused`
3. `TaskSuspender.resume` -> suspender resumed
4. task status is set to `Running`
4. `Paused` event processed -> pause suspender
5. return to p.1

We might get an infinite update loop in this case.

To avoid that we can provide source along with the status change event and filter out events generated by `TaskSuspender` (`SYSTEM`).
In this case we're going to react only on user events (UI button clicks) and p.4-5 won't happen

GitOrigin-RevId: 4fd130fa7a0c270505e0ae13c139211d059a1c3d