Detect how many GC cycles happened between notifications
In JBR 21 MemoryNotificationInfo.MEMORY_COLLECTION_THRESHOLD_EXCEEDED is "broken" and triggered way before major collections happens. As a result, we get increased number of notifications and many memory reports where most objects are soft-referenced.

Instead we're trying to track the number of GC collections for the last minute and if we have 20 of them, it indicates that JVM is struggling under GC.

GitOrigin-RevId: 139128932a7117fe83307569babdde7b4739661e