[debugger] Fix rt.jar selection when debugging IDEA
The idea_rt.jar should always match the IDEA build. However, when running from sources, the corresponding module is a part of the build, so its content was used while debugging, even when the versions do not match. This commit deprioritizes the module in favor of idea_rt.jar from the installation directory.

GitOrigin-RevId: e2aa199936f09485425bdd3e126fc13222f8d409