IJPL-174190 Fix ISE Project is not found for ProjectEntity in `ValueLookupManagerSubscriptions`
The exception might happen when the second project is opened.
`subscribeForDebuggingStart` compares all received `XDebuggerValueLookupListeningStartedEntity` with the provided project.

It might happen that the entity has been created (and received) before the second project on the frontend is initialized. So `projectEntity.asProject()` will throw an exception.

However, we can make a softer check here by calling `asProjectOrNull` - if it returns null, then the project probably hasn't been created yet.
When it is created, `subscribeForDebuggingStart` will be called with an instance of the project and the entity will be processed.

GitOrigin-RevId: 3d4b131584c8d3b31d2513b4374ed30e86c79f31